refactor(desktop): split chatview into sub-files
