+++ apps/desktop/src/features/chat/components/ChatView/index.tsx
@@ -1,29 +1,7 @@
-import {
-  useCallback,
-  useDeferredValue,
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-  type ComponentType,
-} from 'react';
-import { ArrowDown, Sparkles } from 'lucide-react';
-import { DogMascot } from '../../../../shared/components/DogMascot';
-import agentDebugger from '../../../../assets/agents/debugger.png';
-import agentDocs from '../../../../assets/agents/docs.png';
-import agentGoodboy from '../../../../assets/agents/goodboy.png';
-import agentImplementer from '../../../../assets/agents/implementer.png';
-import agentPlanner from '../../../../assets/agents/planner.png';
-import agentReviewer from '../../../../assets/agents/reviewer.png';
-import agentScout from '../../../../assets/agents/scout.png';
-import agentTester from '../../../../assets/agents/tester.png';
-import {
-  AGENT_KIND_META,
-  inferAgentKindFromName,
-  type AgentKind as AgentKindLabel,
-} from '../../../session/agent-kind';
+import { useCallback, useDeferredValue, useEffect, useMemo, useState } from 'react';
+import { ArrowDown } from 'lucide-react';
 import type { AgentId, OpenQuestion, ProviderRunId, Session } from '@goodboy/types';
-import { cn, Divider, Skeleton } from '@goodboy/ui';
+import { cn, Divider } from '@goodboy/ui';
 import {
   EMPTY_ARRAY,
   useAppStore,
@@ -32,11 +10,7 @@ import {
   useSessionOpenQuestions,
   useTranscript,
 } from '../../../../store';
-import {
-  detectParallelRunIds,
-  filterEventsByRunId,
-  reduceTranscript,
-} from '../../utils/transcript-items';
+import { detectParallelRunIds, reduceTranscript } from '../../utils/transcript-items';
 import { clusterOperations } from '../../utils/cluster-operations';
 import { classifyThinkingContext } from '../../utils/thinking-context';
 import { ThinkingIndicator } from '../ThinkingIndicator';
@@ -54,160 +28,18 @@ import {
 import { DiffViewerDialog } from '../../../../features/permissions/components/DiffViewerDialog';
 import { worktreeDiff } from '../../../../features/worktree/worktree';
 import { OpenQuestionCluster } from './OpenQuestionCluster';
+import { ChatEmptyState } from './ChatEmptyState';
+import { ParallelColumn } from './ParallelColumn';
+import { useScrollPin } from './useScrollPin';
+import { dayKey, formatDayLabel } from './lib';
+import { TranscriptSkeleton } from './parts/TranscriptSkeleton';
 
 type ChatViewProps = {
   session: Session;
   isActive?: boolean;
 };
 
-const PIN_TOLERANCE_PX = 32;
 const TERMINAL_STATUSES = new Set(['completed', 'failed', 'cancelled']);
-
-function useScrollPin(deps: ReadonlyArray<unknown>, resetKey?: unknown) {
-  const scrollerRef = useRef<HTMLDivElement>(null);
-  const [pinned, setPinned] = useState(true);
-  const [atTop, setAtTop] = useState(true);
-
-  useEffect(() => {
-    setPinned(true);
-  }, [resetKey]);
-
-  useEffect(() => {
-    const el = scrollerRef.current;
-    if (!el || !pinned) {
-      return;
-    }
-    el.scrollTop = el.scrollHeight;
-  }, [pinned, ...deps]); // eslint-disable-line react-hooks/exhaustive-deps
-
-  const onScroll = () => {
-    const el = scrollerRef.current;
-    if (!el) {
-      return;
-    }
-    const distance = el.scrollHeight - el.scrollTop - el.clientHeight;
-    setPinned(distance < PIN_TOLERANCE_PX);
-    setAtTop(el.scrollTop < PIN_TOLERANCE_PX);
-  };
-
-  return { scrollerRef, pinned, atTop, onScroll };
-}
-
-type ColumnProps = {
-  runId: ProviderRunId;
-  index: number;
-  events: ReturnType<typeof useTranscript>;
-  workingDir: string | null;
-  onRefreshAuth: () => void;
-  onOpenDiff: (filePath: string) => void;
-};
-
-function ParallelColumn({
-  runId,
-  index,
-  events,
-  workingDir,
-  onRefreshAuth,
-  onOpenDiff,
-}: ColumnProps) {
-  const columnEvents = useMemo(() => filterEventsByRunId(events, runId), [events, runId]);
-  const items = useMemo(() => reduceTranscript(columnEvents), [columnEvents]);
-  const { scrollerRef, pinned, onScroll } = useScrollPin([items]);
-
-  return (
-    <div
-      data-run-column={runId}
-      className="flex min-w-0 flex-col border-r border-border last:border-r-0"
-    >
-      <div className="border-b border-border px-3 py-1.5 text-xs font-medium text-muted-foreground">
-        p{index + 1}
-      </div>
-      <div className="relative flex min-h-0 flex-1 flex-col">
-        <div ref={scrollerRef} onScroll={onScroll} className="flex-1 overflow-y-auto px-3 py-3">
-          {items.length === 0 ? (
-            <p className="text-xs text-muted-foreground">no events yet for run p{index + 1}.</p>
-          ) : (
-            <ul className="flex flex-col gap-2.5">
-              {items.map((item) => (
-                <li key={item.key}>
-                  <TranscriptCard
-                    item={item}
-                    workingDir={workingDir}
-                    onRefreshAuth={onRefreshAuth}
-                    onOpenDiff={onOpenDiff}
-                  />
-                </li>
-              ))}
-            </ul>
-          )}
-        </div>
-        {!pinned && (
-          <button
-            type="button"
-            aria-label="jump to latest"
-            title="jump to latest"
-            className="pointer-events-auto absolute bottom-3 left-1/2 z-10 -translate-x-1/2 flex h-8 w-8 items-center justify-center rounded-full border border-border-soft bg-background/90 ring-1 ring-border-soft transition-colors hover:bg-muted"
-            onClick={() => {
-              const el = scrollerRef.current;
-              el?.scrollTo({ top: el.scrollHeight, behavior: 'smooth' });
-            }}
-          >
-            <ArrowDown size={14} aria-hidden />
-          </button>
-        )}
-      </div>
-    </div>
-  );
-}
-
-function dayKey(iso: string): string {
-  const d = new Date(iso);
-  if (Number.isNaN(d.getTime())) {
-    return 'unknown';
-  }
-  return `${d.getFullYear()}-${d.getMonth()}-${d.getDate()}`;
-}
-
-function formatDayLabel(iso: string): string {
-  const d = new Date(iso);
-  if (Number.isNaN(d.getTime())) {
-    return '';
-  }
-  const now = new Date();
-  const startOf = (x: Date) => new Date(x.getFullYear(), x.getMonth(), x.getDate()).getTime();
-  const diffDays = Math.round((startOf(now) - startOf(d)) / 86_400_000);
-  if (diffDays === 0) {
-    return 'today';
-  }
-  if (diffDays === 1) {
-    return 'yesterday';
-  }
-  if (diffDays > 0 && diffDays < 7) {
-    return d.toLocaleDateString(undefined, { weekday: 'long' }).toLowerCase();
-  }
-  return d
-    .toLocaleDateString(undefined, { day: 'numeric', month: 'short', year: 'numeric' })
-    .toLowerCase();
-}
-
-function TranscriptSkeleton() {
-  return (
-    <div
-      role="status"
-      aria-label="loading transcript"
-      className="mx-auto flex w-full max-w-[880px] flex-col gap-6"
-    >
-      {[0, 1, 2].map((i) => (
-        <div key={i} className="flex flex-col gap-2">
-          <Skeleton className="h-3 w-24" />
-          <Skeleton className="h-3 w-full" />
-          <Skeleton className="h-3 w-5/6" />
-          <Skeleton className="h-3 w-3/4" />
-        </div>
-      ))}
-    </div>
-  );
-}
 
 export const ChatView = ({ session, isActive = true }: ChatViewProps) => {
   const selectedAgentId = useAppStore(
@@ -693,194 +525,4 @@ export const ChatView = ({ session, isActive = true }: ChatViewProps) => {
       />
     </div>
   );
-};
-
-type ChatEmptyStateProps = {
-  selectedAgentId: AgentId | null;
-  phaseRuns: ReadonlyArray<import('@goodboy/types').Agent>;
-  hasWorkflow: boolean;
-};
-
-function ChatEmptyState({ selectedAgentId, phaseRuns, hasWorkflow }: ChatEmptyStateProps) {
-  const agentKindOverride = useAppStore((s) => s.agentKindOverride);
-  const selectedAgent = useMemo(
-    () => (selectedAgentId ? (phaseRuns.find((r) => r.id === selectedAgentId) ?? null) : null),
-    [selectedAgentId, phaseRuns],
-  );
-  const selectedKind = useMemo(() => {
-    if (!selectedAgent) {
-      return null;
-    }
-    return agentKindOverride[selectedAgent.id] ?? inferAgentKindFromName(selectedAgent.name);
-  }, [selectedAgent, agentKindOverride]);
-
-  const scenario = useMemo<EmptyScenario>(() => {
-    if (selectedAgent && selectedKind) {
-      return 'agent_focus';
-    }
-    if (phaseRuns.length > 0) {
-      return 'pick_agent';
-    }
-    if (hasWorkflow) {
-      return 'workflow_no_agent';
-    }
-    return 'fresh';
-  }, [selectedAgent, selectedKind, phaseRuns.length, hasWorkflow]);
-
-  const copy = useMemo<EmptyCopy>(() => {
-    switch (scenario) {
-      case 'agent_focus': {
-        const meta = AGENT_KIND_META[selectedKind as AgentKindLabel];
-        return {
-          eyebrow: `${meta.label} agent · fresh transcript`,
-          title: `You're talking to a ${meta.label} agent`,
-          body: `${meta.hint}. It already knows the session brief on the right: goal, decisions, open questions. No need to re-explain. Say what you want next.`,
-          hints: [
-            selectedKind === 'scout' ? 'Try: "find where X is defined"' : null,
-            selectedKind === 'planner' ? 'Try: "plan how to add X to Y"' : null,
-            selectedKind === 'implementer' ? 'Try: "implement step 2 of the plan"' : null,
-            selectedKind === 'debugger' ? 'Try: "reproduce: <stack trace>"' : null,
-            selectedKind === 'tester' ? 'Try: "write tests for X"' : null,
-            selectedKind === 'reviewer' ? 'Try: "review the current diff"' : null,
-            selectedKind === 'resolver' ? 'Spawned automatically by the resolve UI.' : null,
-            '⌘↵ to send',
-          ].filter((x): x is string => Boolean(x)),
-        };
-      }
-      case 'pick_agent':
-        return {
-          eyebrow: `${phaseRuns.length === 1 ? 'agent' : 'agents'} in this session`,
-          title: 'Pick an agent on the left',
-          body: 'Agents share the session context on the right. Every new one starts already knowing the goal, decisions and open questions. Only the chat history is per-agent. Pick one to keep talking, or spawn a new one: it will hit the ground running.',
-          hints: ['Select an agent to see its transcript', 'Spawn fresh, context travels with it'],
-        };
-      case 'workflow_no_agent':
-        return {
-          eyebrow: 'Workflow ready · No agents yet',
-          title: 'Start the first step',
-          body: 'No agents have run yet. Write here to shape the session brief on the right: goal, constraints, anything important. The first agent, and every one after, will start already knowing it.',
-          hints: ['Describe the goal in 1–2 lines', 'Lands in the shared context'],
-        };
-      case 'fresh':
-      default:
-        return {
-          eyebrow: 'Fresh session · No context yet',
-          title: "Let's populate the context",
-          body: "Whatever you write here feeds the shared session brief on the right: goal, decisions, open questions. Every agent you spawn from now on starts already knowing the essentials, so you don't repeat yourself.",
-          hints: [
-            'What are we building',
-            'Any constraints or non-goals',
-            'Who should the first agent be',
-          ],
-        };
-    }
-  }, [scenario, selectedKind, phaseRuns.length]);
-
-  const agentVisual = scenario === 'agent_focus' && selectedKind ? KIND_ICON[selectedKind] : null;
-
-  return (
-    <div className="mx-auto flex w-full max-w-[640px] flex-col items-center justify-center gap-5 px-6 py-16 text-center">
-      <div className="flex items-center justify-center">
-        {agentVisual?.image ? (
-          <MaskedDog image={agentVisual.image} className={cn('size-32', agentVisual.tint)} />
-        ) : scenario === 'pick_agent' ? (
-          <span className="text-7xl font-semibold leading-none tracking-tight tabular-nums text-foreground">
-            {phaseRuns.length}
-          </span>
-        ) : (
-          <DogMascot size={128} className="text-primary" />
-        )}
-      </div>
-      <div className="flex flex-col gap-1.5">
-        <span className="text-2xs font-semibold uppercase tracking-[0.12em] text-muted-foreground/70">
-          {copy.eyebrow}
-        </span>
-        <h2 className="text-base font-semibold text-foreground">{copy.title}</h2>
-        <p className="text-sm leading-relaxed text-muted-foreground">{copy.body}</p>
-      </div>
-      <div className="flex flex-wrap items-center justify-center gap-1.5 text-2xs text-muted-foreground/70">
-        {copy.hints.map((hint) => (
-          <span
-            key={hint}
-            className="inline-flex items-center gap-1 rounded-full border border-border-soft bg-background px-2 py-0.5 text-2xs"
-          >
-            <Sparkles size={10} aria-hidden />
-            {hint}
-          </span>
-        ))}
-      </div>
-    </div>
-  );
-}
-
-function MaskedDog({ image, className }: { image: string; className?: string }) {
-  return (
-    <span
-      aria-hidden
-      className={cn('shrink-0', className)}
-      style={{
-        maskImage: `url(${image})`,
-        maskRepeat: 'no-repeat',
-        maskPosition: 'center',
-        maskSize: 'contain',
-        WebkitMaskImage: `url(${image})`,
-        WebkitMaskRepeat: 'no-repeat',
-        WebkitMaskPosition: 'center',
-        WebkitMaskSize: 'contain',
-      }}
-    />
-  );
-}
-
-type EmptyScenario = 'fresh' | 'workflow_no_agent' | 'pick_agent' | 'agent_focus';
-
-type KindVisual = {
-  image: string | null;
-  tint: string;
-};
-
-const KIND_ICON: Record<AgentKindLabel, KindVisual> = {
-  generic: {
-    image: agentGoodboy,
-    tint: 'bg-rose-400',
-  },
-  scout: {
-    image: agentScout,
-    tint: 'bg-sky-400',
-  },
-  planner: {
-    image: agentPlanner,
-    tint: 'bg-violet-400',
-  },
-  implementer: {
-    image: agentImplementer,
-    tint: 'bg-emerald-400',
-  },
-  debugger: {
-    image: agentDebugger,
-    tint: 'bg-amber-400',
-  },
-  tester: {
-    image: agentTester,
-    tint: 'bg-teal-400',
-  },
-  reviewer: {
-    image: agentReviewer,
-    tint: 'bg-cyan-400',
-  },
-  docs: {
-    image: agentDocs,
-    tint: 'bg-orange-400',
-  },
-  resolver: {
-    image: null,
-    tint: 'bg-lime-400',
-  },
-};
-
-type EmptyCopy = {
-  eyebrow: string;
-  title: string;
-  body: string;
-  hints: ReadonlyArray<string>;
 };
